+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 
 ## [v1.45.3] - 2024-06-05
 
+### Added
+- Add image cropper to setting group avatars
+
 ### Fix
 - Fix map in packaged build #3900
+- Fix removing group avatars
 
 <a id="1_45_2"></a>